chore: moved assets to public
